FISH-316 Change New Relic Notifier Package
