+++ appserver/payara-appserver-modules/notification-newrelic-core/src/main/java/fish/payara/extras/notifiers/newrelic/GetNewRelicNotifierConfigurationCommand.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.notification.newrelic;
+package fish.payara.extras.notifiers.newrelic;
 
 import java.util.Map;
