+++ appserver/payara-appserver-modules/notification-newrelic-core/src/main/java/fish/payara/extras/notifiers/newrelic/NewRelicNotifierService.java
@@ -37,9 +37,8 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.notification.newrelic;
+package fish.payara.extras.notifiers.newrelic;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.StringWriter;
