+++ appserver/payara-appserver-modules/notification-newrelic-core/src/main/java/fish/payara/extras/notifiers/newrelic/NewRelicSerializer.java
@@ -1,4 +1,4 @@
-package fish.payara.notification.newrelic;
+package fish.payara.extras.notifiers.newrelic;
 
 import java.io.IOException;
